Reconfigured app to get JS working properly
