--- app/javascript/packs/index.js
@@ -1,21 +1,0 @@
-import React from 'react'
-import { render } from 'react-dom'
-import RedBox from 'redbox-react'
-import App from '../components/App'
-
-document.addEventListener('DOMContentLoaded', () => {
-  let reactElement = document.getElementById('app')
-
-  if (reactElement) {
-    if (window.railsEnv && window.railsEnv === 'development') {
-      try {
-        render(<App />, reactElement)
-      } catch (e) {
-        render(<RedBox error={e} />, reactElement)
-      }
-    }
-    else {
-      render(<App />, reactElement)
-    }
-  }
-})
