+++ app/views/layouts/application.html.erb
@@ -16,9 +16,7 @@
         <a class="navbar-item logo" href="/">
         <img src="/images/logo.png">
       </a>
-      <a class="navbar-item logo-name" href="/">
-        <img src="/images/logo-name.png">
-      </a>
+      
 
         <a role="button" class="navbar-burger" aria-label="menu" aria-expanded="false" data-target="navbarBasicExample">
           <span aria-hidden="true"></span>
@@ -65,13 +63,16 @@
       <h4 class="is-size-4">&copy; Perpetually Lost Photography</h4>
     </footer>
     
-    <%- if Rails.env.development? || Rails.env.test? -%> 
-      <%= javascript_tag do %>
-        window.railsEnv = '<%= Rails.env %>'
-      <%- end -%>
+    <%= javascript_include_tag 'application' %> 
+  
+  <%- if Rails.env.development? || Rails.env.test? -%> 
+    <%= javascript_tag do %>
+      window.railsEnv = '<%= Rails.env %>'
     <%- end -%>
-    
-    <%= javascript_pack_tag 'index' %>
+  <%- end -%>
+  <%= javascript_pack_tag 'application' %>
+  
+  <%= yield :extra_footer %>
 
   </body>
 </html>
